Rename plugin to welld-kogito-kjar-maven-plugin
