+++ .github/workflows/feature.yml
@@ -2,7 +2,7 @@ name: Feature workflow
 on:
   push:
     branches: 
-    - '!master'
+    - 'feature/**'
 
 env:
   MAVEN_CLI_OPTS: "-s .m2/settings.xml --batch-mode"
